# docs/account/swagger-spec.json
{"openapi":"3.0.0","paths":{"/users/me":{"get":{"operationId":"UserController_getMyProfile","summary":"내 프로필 조회","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"tags":["Account","User"],"security":[{"access-token":[]}]},"post":{"operationId":"UserController_createMyProfile","summary":"내 프로필 생성","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserBody"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"tags":["Account","User"],"security":[{"access-token":[]}]}},"/users/me/{id}":{"put":{"operationId":"UserController_updateUser","summary":"내 프로필 수정","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserBody"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"tags":["Account","User"],"security":[{"access-token":[]}]}},"/users/me/images/{no}":{"post":{"operationId":"UserController_getPresignedUrl","summary":"프로필 이미지 추가/수정을 위한 PresignedUrl 생성","parameters":[{"name":"no","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPresignedUrlResponse"}}}}},"tags":["Account","User"],"security":[{"access-token":[]}]},"delete":{"operationId":"UserController_deletePresignedUrl","summary":"프로필 이미지 삭제","parameters":[],"responses":{"204":{"description":"이미지 삭제 성공"}},"tags":["Account","User"],"security":[{"access-token":[]}]}},"/users/{id}":{"get":{"operationId":"UserController_getUser","summary":"특정 사용자 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"tags":["Account","User"]}},"/users":{"delete":{"operationId":"UserController_deleteUser","summary":"내 계정 삭제","parameters":[],"responses":{"204":{"description":"계정 삭제 성공"}},"tags":["Account","User"],"security":[{"access-token":[]}]}},"/auth/{provider}":{"get":{"operationId":"AuthController_getSocialLoginPage","summary":"소셜 로그인 페이지 요청","parameters":[{"name":"provider","required":true,"in":"path","description":"소셜 로그인 제공자","schema":{"enum":["KAKAO","NAVER","GOOGLE","APPLE","EMAIL","from"],"type":"string"}}],"responses":{"302":{"description":"소셜로그인 페이지 리다이렉트"}},"tags":["Account","Auth"]}},"/auth/{provider}/callback":{"post":{"operationId":"AuthController_socialCallback","summary":"소셜 로그인 콜백 처리 (POST 방식)","parameters":[{"name":"provider","required":true,"in":"path","description":"소셜 로그인 제공자","schema":{"enum":["KAKAO","NAVER","GOOGLE","APPLE","EMAIL","from"],"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SocialCallbackDto"}}}},"responses":{"200":{"description":"소셜로그인 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"tags":["Account","Auth"]},"get":{"operationId":"AuthController_socialCallbackGet","summary":"소셜 로그인 콜백 처리 (GET 방식)","parameters":[{"name":"provider","required":true,"in":"path","description":"소셜 로그인 제공자","schema":{"enum":["KAKAO","NAVER","GOOGLE","APPLE","EMAIL","from"],"type":"string"}},{"name":"code","required":true,"in":"query","description":"인증 코드","schema":{"type":"string"}},{"name":"state","required":false,"in":"query","description":"상태 값(선택)","schema":{"type":"string"}}],"responses":{"200":{"description":"소셜로그인 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"tags":["Account","Auth"]}},"/auth/principal/{providerType}/{providerId}":{"get":{"operationId":"AuthController_getUserPrincipal","summary":"유저 정보 획득","description":"해당 API 는 각 서비스에서 유저정보를 획득하기 위해 사용합니다. 클라이언트에서 사용하지 않습니다.","parameters":[{"name":"providerType","required":true,"in":"path","description":"소셜 로그인 제공자","schema":{"enum":["KAKAO","NAVER","GOOGLE","APPLE","EMAIL","from"],"type":"string"}},{"name":"providerId","required":true,"in":"path","description":"소셜 로그인 제공자 ID","schema":{"type":"string"}}],"responses":{"200":{"description":"유저 정보 획득 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserPrincipalResponse"}}}}},"tags":["Account","Auth"]}},"/auth/{provider}/native":{"post":{"operationId":"AuthController_socialLoginNative","summary":"소셜 로그인 (네이티브 방식)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SocialLoginNativeBody"}}}},"responses":{"200":{"description":"소셜로그인 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"tags":["Account","Auth"]}},"/auth/test/{provider}":{"get":{"operationId":"AuthController_testSocialLogin","parameters":[{"name":"provider","required":true,"in":"path","description":"소셜 로그인 제공자","schema":{"enum":["KAKAO","NAVER","GOOGLE","APPLE","EMAIL","from"],"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Account","Auth"]}}},"info":{"title":"개개팅 API","description":"강아지 산책 소개팅 애플리케이션 API 문서","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"JWT 토큰을 입력하세요","in":"header"}},"schemas":{"UserResponse":{"type":"object","properties":{"id":{"type":"string","description":"사용자 ID"},"nickname":{"type":"string","description":"닉네임"},"gender":{"type":"string","description":"성별"},"region":{"type":"string","description":"지역"},"bio":{"type":"string","description":"자기소개"},"phoneNumber":{"type":"string","description":"전화번호"},"profileImages":{"description":"프로필이미지","type":"array","items":{"type":"string"}}},"required":["id","nickname","gender","region","bio","phoneNumber","profileImages"]},"CreateUserBody":{"type":"object","properties":{"name":{"type":"string","description":"이름"},"email":{"type":"string","description":"이메일 주소"},"password":{"type":"string","description":"비밀번호"},"nickname":{"type":"string","description":"닉네임"},"gender":{"type":"string","description":"성별","enum":["MALE","FEMALE","from"]},"birthDate":{"format":"date-time","type":"string","description":"생년월일 (YYYY-MM-DD 형식)","example":"2000-01-01"},"region":{"type":"string","description":"지역","enum":["SEOUL","GYEONGGI","INCHEON","GANGWON","CHUNGCHEONG","JEOLLA","GYEONGSANG","JEJU","from"]},"bio":{"type":"string","description":"자기소개"},"phoneNumber":{"type":"string","description":"전화번호"}},"required":["name","nickname","gender","birthDate","region"]},"UpdateUserBody":{"type":"object","properties":{"nickname":{"type":"string","description":"닉네임"},"gender":{"type":"string","description":"성별","enum":["MALE","FEMALE","from"]},"region":{"type":"string","description":"지역","enum":["SEOUL","GYEONGGI","INCHEON","GANGWON","CHUNGCHEONG","JEOLLA","GYEONGSANG","JEJU","from"]},"bio":{"type":"string","description":"자기소개"}},"required":["gender","region"]},"GetPresignedUrlResponse":{"type":"object","properties":{"url":{"type":"string","description":"사진 업로드 URL"},"expiresIn":{"type":"number","description":"URL 만료 시간(초)"}},"required":["url","expiresIn"]},"SocialCallbackDto":{"type":"object","properties":{"code":{"type":"string","description":"인증 코드"},"state":{"type":"string","description":"상태 값(선택)"}},"required":["code"]},"LoginResponse":{"type":"object","properties":{"accessToken":{"type":"string","description":"액세스 토큰"},"expiresIn":{"type":"number","description":"액세스 토큰 만료 시간(초)"},"refreshToken":{"type":"string","description":"리프레시 토큰"},"refreshTokenExpiresIn":{"type":"number","description":"리프레시 토큰 만료 시간(초)"}},"required":["accessToken","expiresIn","refreshToken","refreshTokenExpiresIn"]},"GetUserPrincipalResponse":{"type":"object","properties":{"userId":{"type":"string","description":"사용자ID"},"nickname":{"type":"string","description":"닉네임"},"birth":{"type":"string","description":"생년월일"},"region":{"type":"number","description":"거주지번호"}},"required":["userId","nickname","birth","region"]},"SocialLoginNativeBody":{"type":"object","properties":{"accessToken":{"type":"string","description":"소셜에서 발급받은 액세스토큰"},"refreshToken":{"type":"string","description":"소셜에서 발급받은 리프레시토큰"},"expiresIn":{"type":"number","description":"액세스 토큰 만료 시간 (초)"},"refreshTokenExpiresIn":{"type":"number","description":"리프레시 토큰 만료 시간 (초)"},"tokenType":{"type":"string","description":"액세스 토큰 타입"}},"required":["accessToken","refreshToken","expiresIn","refreshTokenExpiresIn","tokenType"]}}}}